Changed - Various core files in preparation for GEDFact_assistant module
git-svn-id: svn://svn.webtrees.net/trunk/phpGedView@5491 4201c862-f99b-1943-94d0-25fd05010cbc

+++ inverselink.php
@@ -119,14 +119,30 @@ if ($action == "choose" && $paramok) {
  	} else {
 		echo '<input type="text" name="mediaid" id="mediaid" size="5" />';
 		print_findmedia_link("mediaid","1media");
+		echo "</td></tr>";
 	}
-	echo '</td></tr>';
+	
+	// GEDFact assistant Current Media Links ===================
+	if (file_exists('modules/GEDFact_assistant/MEDIA/media_1_ctrl.php')) {
+		echo "<tr>";
+		echo "<td class=\"descriptionbox width20 wrap\">";
+		echo "Current links:";
+		echo"</td>";
+		echo "<td class=\"optionbox wrap\">";
+			include ('modules/GEDFact_assistant/MEDIA/media_query_1a.php');
+		echo "</td></tr>";
+	}
+	// =========================================================
+	
 
 	if (!isset($linktoid)) $linktoid = "";
 	print "<tr><td class=\"descriptionbox\">";
+	
 	if ($linkto == "person") {
-		print $pgv_lang["enter_pid"]."</td>";
+		// print $pgv_lang["enter_pid"]."</td>";
+		echo "Add more links:";
 		print "<td class=\"optionbox wrap\">";
+		echo "First, enter a Base individual ID<br />";
 		if ($linktoid=="") {
 			print "<input class=\"pedigree_form\" type=\"text\" name=\"linktoid\" id=\"linktopid\" size=\"3\" value=\"$linktoid\" />";
 			print_findindi_link("linktopid","");
@@ -137,6 +153,21 @@ if ($action == "choose" && $paramok) {
 			print "(".$linktoid.")";
 			if ($TEXT_DIRECTION=="rtl") print getRLM();
 		}
+		echo "<br /><br />";
+		echo "Then, click Add to add more Individual Links";
+		echo "<br />";
+		
+		// GEDFact assistant Add Media Links =======================
+		if (file_exists('modules/GEDFact_assistant/MEDIA/media_1_ctrl.php')) {
+			include ('modules/GEDFact_assistant/MEDIA/media_query_2a.php');
+			echo "</td></tr>";
+			echo "<tr><td class=\"topbottombar\" colspan=\"2\">";
+			echo "<input type=\"button\" value=\"".$pgv_lang["set_link"]."s\" onclick=\"javascript:alert('Clicking \'Set Links\' will eventually parse and save the Current and Added Links');\" />";
+			echo "</td></tr>";
+		}else{
+		print "<tr><td class=\"topbottombar\" colspan=\"2\"><input type=\"submit\" value=\"".$pgv_lang["set_link"]."\" /></td></tr>";
+		}
+
 	}
 
 	if ($linkto == "family") {
@@ -152,6 +183,8 @@ if ($action == "choose" && $paramok) {
 			print "(".$linktoid.")";
 			if ($TEXT_DIRECTION=="rtl") print getRLM();
 		}
+		print "</td></tr>";
+		print "<tr><td class=\"topbottombar\" colspan=\"2\"><input type=\"submit\" value=\"".$pgv_lang["set_link"]."\" /></td></tr>";
 	}
 
 	if ($linkto == "source") {
@@ -167,9 +200,13 @@ if ($action == "choose" && $paramok) {
 			print "(".$linktoid.")";
 			if ($TEXT_DIRECTION=="rtl") print getRLM();
 		}
+		print "</td></tr>";
+		print "<tr><td class=\"topbottombar\" colspan=\"2\"><input type=\"submit\" value=\"".$pgv_lang["set_link"]."\" /></td></tr>";
 	}
-	print "</td></tr>";
-	print "<tr><td class=\"topbottombar\" colspan=\"2\"><input type=\"submit\" value=\"".$pgv_lang["set_link"]."\" /></td></tr>";
+
+	
+	
+
 	print "</table>";
 	print "</form>\n";
 	print "<br/><br/><center><a href=\"javascript:;\" onclick=\"if (window.opener.showchanges) window.opener.showchanges(); window.close();\">".$pgv_lang["close_window"]."</a><br /></center>\n";
